pyproject.toml
[tool.poetry]
name = "musicinfo"
version = "0.1.10"
description = ""
authors = ["Michael Huahuasoncco <michael_h_p_@hotmail.com>"]
readme = "README.md"
packages = [{include = "musicinfo"}]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.3.5"
langchain-openai = "^0.3.0"
langchain-core = "^0.3.10"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.13.3"
heyoo = "^0.1.2"
langchain-deepseek = "^0.1.2"

[tool.poetry.group.dev.dependencies]
langgraph-cli = "^0.1.77"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env"